icuregex: Lazy load ICU data into memory
This avoids loading the ICU data into memory for parts of Vitess that
never end up using it, like `vtorc` or `vttablet`.
